move assertions to a single file
